working on order management arrays
