--- src/components/OrderManagement/Foodlist.js
@@ -1,22 +1,0 @@
-import amala from './Images/' 
-
-
-const FoodList = [
-    {
-        id: 1,
-        logo: "HA",
-        name: "Hafsat Abdul",
-        orderCode: 24071,
-        date: "12-08-2023",
-        time: "11:28 PM",
-        address: "49 Featherstone Street, EC1Y 8SY, London",
-        Foodname1: "Amala & Ewedu",
-        Desc1: "Amala, Ewedu and Gbegiri",
-        img1: "./Images/",
-        Price1: 20,
-        Foodname2: "Fried Rice & Chicken",
-        Desc2: "Fried Rice, Salad and Chicken",
-        img2: "./Images/",
-        Price2: 18,
-    }
-]
