+++ src/lib/Foodlist.js
@@ -1,0 +1,100 @@
+import Amala from '../assets/Images-Order-management/Amala&Ewedu.jpg'; 
+import BeansPlantain from '../assets/Images-Order-management/Beans&Plantain.jpg';
+import Egusi from '../assets/Images-Order-management/Egusi&Poundo.jpg'; 
+import EwaAgoyin from '../assets/Images-Order-management/EwaAgoyin.jpg'; 
+import FriedRice from '../assets/Images-Order-management/Friedrice.jpg'; 
+import GhanaJollof from '../assets/Images-Order-management/GhanaJollof.jpg'; 
+import Oha from '../assets/Images-Order-management/Oha.jpg'; 
+import Waakye from '../assets/Images-Order-management/Waakye.jpg'; 
+
+
+const FoodList = [
+    {
+        id: 1,
+        logo: "HA",
+        name: "Hafsat Abdul",
+        orderCode: 24071,
+        date: "12-08-2023",
+        time: "11:28 PM",
+        address: "49 Featherstone Street, EC1Y 8SY, London",
+        Foodname1: "Amala & Ewedu",
+        Desc1: "Amala, Ewedu and Gbegiri",
+        img1: Amala,
+        Price1: 20,
+        Qty1: 2,
+        Foodname2: "Fried Rice & Chicken",
+        Desc2: "Fried Rice, Salad and Chicken",
+        img2: FriedRice,
+        Price2: 18,
+        Qty2: 2,
+        // Total: 38,
+    },
+
+
+    {
+        id: 2,
+        logo: "HA",
+        name: "Hafsat Abdul",
+        orderCode: 24071,
+        date: "12-08-2023",
+        time: "11:28 PM",
+        address: "49 Featherstone Street, EC1Y 8SY, London",
+        Foodname1: "Amala & Ewedu",
+        Desc1: "Amala, Ewedu and Gbegiri",
+        img1: Amala,
+        Price1: 20,
+        Qty1: 2,
+        Foodname2: "Fried Rice & Chicken",
+        Desc2: "Fried Rice, Salad and Chicken",
+        img2: FriedRice,
+        Price2: 18,
+        Qty2: 2,
+        // Total: 38,
+    },
+
+
+    {
+        id: 3,
+        logo: "HA",
+        name: "Hafsat Abdul",
+        orderCode: 24071,
+        date: "12-08-2023",
+        time: "11:28 PM",
+        address: "49 Featherstone Street, EC1Y 8SY, London",
+        Foodname1: "Amala & Ewedu",
+        Desc1: "Amala, Ewedu and Gbegiri",
+        img1: Amala,
+        Price1: 20,
+        Qty1: 2,
+        Foodname2: "Fried Rice & Chicken",
+        Desc2: "Fried Rice, Salad and Chicken",
+        img2: FriedRice,
+        Price2: 18,
+        Qty2: 2,
+        // Total: 38,
+    },
+
+
+    {
+        id: 4,
+        logo: "HA",
+        name: "Hafsat Abdul",
+        orderCode: 24071,
+        date: "12-08-2023",
+        time: "11:28 PM",
+        address: "49 Featherstone Street, EC1Y 8SY, London",
+        Foodname1: "Amala & Ewedu",
+        Desc1: "Amala, Ewedu and Gbegiri",
+        img1: Amala,
+        Price1: 20,
+        Qty1: 2,
+        Foodname2: "Fried Rice & Chicken",
+        Desc2: "Fried Rice, Salad and Chicken",
+        img2: FriedRice,
+        Price2: 18,
+        Qty2: 2,
+        // Total: 38,
+    }
+]
+
+export default FoodList;
